requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.7.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
filelock==3.19.1
fsspec==2025.9.0
gunicorn==23.0.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.8.2
